[WINDOWS] Get tracing in the new messaging setup working under windows again.
Windows does not support Domain sockets so for windows it should be possible to
define a port number on the loopback interface (127.0.0.1) as a domain socket alternative.
This pull request adds the possibility (if the port is not 0) to use a loopback IP socket
for communication in stead of a domain socket. This allwos the messaging concept to also
work on a windows platform.