[WINDOWS] Get tracing in the new messaging setup working under window… (#950)
* [WINDOWS] Get tracing in the new messaging setup working under windows again.

Windows does not support Domain sockets so for windows it should be possible to
define a port number on the loopback interface (127.0.0.1) as a domain socket alternative.
This pull request adds the possibility (if the port is not 0) to use a loopback IP socket
for communication in stead of a domain socket. This allwos the messaging concept to also
work on a windows platform.

* Update MessageDispatcher.h

Added the doxygen description for the new parameter.

* Update MessageUnit.cpp

Described the parameter in the docygen section of the open

* Update MessageClient.cpp

Updated the docygen for the new parameter

* Update MessageDispatcher.h

Create a default for the new parameter, so the default allows for working with domain sockets.

* Update MessageUnit.h

On the open, pass a default paramater of 0 in case the usage is only with domain sockets.

* Update MessageClient.cpp

Make sure the default usage of the classes is to use domain sockets in stead of IP sockets.

* Update MessageClient.cpp

Oops wrong spot :-)

* Update MessageClient.h

Make sure the default is to use domain sockets in stead of IP sockets.