Do not lock during destrcution :-)
However during the deep dive into this, I wonder if the real problem was the _destructor. Up till you changes the _destructor member variable was never set to a nullptr and during construction it would even ASSERT if you would pass in a nullptr. So technically the _destructor member variable could never have been a nullptr. Where did you experience this problem. Is there a reproduction scenario ?