RDKTV-15888,RDKTV-15391,RDKTV-15673: Unregister when Socket Closed
Reason for change: When the socket is closed, the object pointer
is not removed from the ResourceMonitor. This is leading to dereference
the pointer that got destructed but not removed from ResourceMonitor
when there is race between closing the socket and some pending event
to be performed on the socket.
Test Procedure: As per RDKTV-15888
Risks: Medium

Signed-off-by:Zameerun Rasheed M S <zexpen@gmail.com>