RDKTV-16992: Added timeout and synchronisation when stopping containers
Reason for change: It was possible for successive calls to DobbyContainer::Stop
to be made due to the WPEFramework not being notified quickly enough
that a container had stopped.
This change implements the missing timeout functionality
within that method to prevent the opportunity for quick multiple calls.
When the timeout is zero this method behaves as previous.

Test Procedure: Stop containers with zero and non-zero timeouts and
ensure there are no occurrences of containers being stopped multiple
times.

Risks: Low

Signed-off-by: Will Tallentire <will.tallentire@sky.uk>