Thread: Fix handling of pthread_create() failures
When pthread_create() fails, it returns a value different than zero, not
-1 as the current implementation was assuming. This was detected in a
use case where some threads were trying to be created with approximately
1GiB stack size and failing.

Additionally, when the thread creation fails, any waits or set states on
it must fail, otherwise the system may become deadlocked waiting for
some thread that could not even be created. For this, introduce a new
'FAILED' state and handle it accordingly.

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>