[COM] Add a class to retrieve the default searchpaths, used by the dl.
Mod dl implementations use the /etc/ld.so.conf to extend the default set of /usr/lib and /lib. This commit
makes a list of default search paths used by the dl to load the dynamic libraries. It can than be used in
the spawning of an out-of-process plugin to search for the libraries in the downloaded areas as well for
that plugin, using the LD_LIBRARAY_PATH (fileld with this retrieved set and prefixes with the "SystemRootPath"
for the plugin.