[SMARTINTERFACETYPE] Fix crash at closing doen the process.
If all singeltons are destructed but there are still some Sockets subscribed to the Monitor,
communication thread, the code is written in a defensive way to notify the developper to
properly cleanup. If the SmartInterfaceType was used, this notification was fire but the
developepr did all in his power to clean up properly ;-) There was a bug in the framework.

This commit resolves this bug. The channel did not act in the unlink() event. The unlink
event is used to signal that the object is no longer being used in the ProxyMap and thus
is nolonger used for communication (in this setup), so the unlink should effectively "close"
the connection. That is what this commit realizes.

As this was testsed and debugged uner windows another bug, only applicable to windows was
found. The proxyStubs that used to be part of the plugins.so (in the past) have been moved
in their own proxystub librray. That was fully completed for linux, but for windows there
was not yet a project file to build it. Now there is :-)