[IDLE] Introduce the concept of Idle reporting on the ThreadPool.
From static ocde analyses and core dumps sporadically received from the field, it was consluded
that dropping/unloading the plugin code from memory at the time the plugin is deactivated was
leading to segmentation fault in WPEFramework as potentially there could be a pending COMRPC
waiting in the WorkerPool to be executed. If the code, triggered from the COMRPC call, hard
casted and calling an address part ot the COMRPC message, could be located in the unloaded space
a segmentation fault could occure.

By adding an Idle notification callback to the ThreadPool (and forwarded into the WorkerPool),
the unloading of the libraries could be triggered from the notification that all potential
COMRPC/JSONRPC work on a plugin from the workerpool is done and thus there is no more
COMRPC/JSONRPC work pending. So now it is safe to unload the library triggered for unloading
during the plugin deactivation.