[JSON] Enhance the JSON parsing to allow for empty array serialization. (#1112)
Using the new method Set on the JSON::ArrayType<> template, the array can be set to Serialize although there are
no elements available in the array. The Serialization will show like: "fieldName": []
This can be used to indicate that a field has been set but there is no data for this field.