Controller: Allow configuring if the UI is enabled or not
Currently, the Controller's UI, serving as base for the Thunder UI, is
enabled by default, without allowing a configuration option to disable
it.

Since having the Controller's UI enabled requires Web Servicing
(accepting and responding to HTTP requests) and that might bring
security concerns, and also considering that the UI is mostly for
development/debug, add a way to configure the Controller to have the UI
disabled.

This way, add a cmake boolean option ('PLUGIN_CONTROLLER_UI_ENABLED')
which, when set to 'false', disables the Controller's UI. Set the new
option as 'true' by default, enabling the UI by default, for keeping
backwards compatibility.

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>