Sync json parser with master (#1168)
* [JSONRPC] Interface methods should not be private or protected of they are public on the interface. (#1058)

* [JSON] Parsing error incorrectly set on none quoted simple types. (#1094)

* [JSON] Parsing error incorrectly set on none quoted simple types.

In the JSONRPC message, one can return "result": 0 to indicate that the registartion request iver JSONRPC succeeded, erro code0. However before
this fix, the result was expected to be quoted, an array or object. No we also assume that plain results are oke.

* [OOPS] Such a small fix such a serious issue, hopefully fixed :-)

* [JSON] Enhance the JSON parsing to allow for empty array serialization. (#1112)

Using the new method Set on the JSON::ArrayType<> template, the array can be set to Serialize although there are
no elements available in the array. The Serialization will show like: "fieldName": []
This can be used to indicate that a field has been set but there is no data for this field.

* [Core][JSON] Add conversion operator to string for JSON::String

* [PARSING] If you quote a string escape thin quotes in the string :-) (#1163)

Co-authored-by: Pierre Wielders <pierre@wielders.net>
Co-authored-by: sebaszm <s.szczepaniak@metrological.com>