Logging: Add module name to logs
Current implementation of the SysLog function won't output the
origin of the log.

With this commit, the resulting log message will have the form:
[<time>]:[<module-name>]:[<file-name:line-number>] <log-category>:
<data>

This would allow for easier handling of log messages in a scenario where
multiple plugins have their tracing flag enabled at the same time.

To capture syslog messages for a given plugin, use `journalctl |grep
<module-name>`.