[WORKERPOOL] Next to the running jobs, export the queued jobs. (#1209)
* [WORKERPOOL] Next to the running jobs, export the queued jobs.

If the Thunder system seems stop procesing requests it is interesting to know what
is actually running, what is deadlocking in the workerpool. Each Job already had a
method to return an Identifier() for the job. For COMRPC and JSONRPC, some elements
of the job were already made part of this identifier, so that for PostMortem analyses
one could deduce the type of job running (which interface, which method etc). This
functionality has now been extended by also reporting what is in the queue (waiting
jobs for the workerpool).
Also if the Workerpool seems/is deadlocked, this information can not be retrieved
through calls that are handled in the workerpool hence why the reporting functionality
is now also exposed through a signal handler on Thuner. If a SIGUSR1 is send all
relevant internal Thunder information (like the WorkerPool jobs) are dumped to
<PostMortemPath>/ThunderInternals.json.

This functionality was triggered by Madana Gopal Thirumalai, this pull-request is
extending his implementation beyond the JSONRPC jobs.

* Make sure the Console output understands the new Pending