[JSONRPC] Resolving the Technical debt of the quickly build JSONRPC i… (#1235)
* [JSONRPC] Resolving the Technical debt of the quickly build JSONRPC interface.

The JSONRPC interface, although designed as COMRPC, did cary objects (Proxies). Since new
features require the JSONRPC to be ran over COMRPC (to avoid multple serialization/
deserializations). THis required to pay the technical debt of the shortcut we took in the
past. This commit makes the JSONRPC::IDispatch interface a "real" COMRPC interface.

* Compilation issue fix

* [FIXES and CLEANUP] aAfter testing..

---------

Co-authored-by: MFransen69 <39826971+MFransen69@users.noreply.github.com>