Development/extensions (#1250)
* [RESTRUCTURE] Move supporting/extension libraries in a separate directory.

As Thunder is groing, more and more supportin (not essential) libraries are added to
the framework. These libraries are not essential to the working of Thunder, they just
add functionality on request. Thunder can build and run without the availability of
any of these libraries..

hibernate:         If enabled allows out of process plugins to be swapped to disk and free up the
                   allocated memory. Usefull on small memory footprint devices.
localtracer:       Usefull header only library for third-party developpers to read/print TRACE
                   statements sed in the code without actally running the Thudner framework.
privilegedrequest: Header only librray, only available for linux. Allows filedescriptors available
                   in one process to be handed over to another process. This library simplifies
                   the boiler plated code to do this, relying on implementation in Thunder Core!
processcontainer:  If the out of process plugin needs to be running in a container, this is the
                   library you need. It abstracts several container implementations (CRUN/RUNC/LXC
                   and Dobby). It allows the lifetime management of THunder and it plugins to
                   seemlessly integrate with the containers.
warningreporting:  If you enable this feature, certain checks in Thunder are activated and Thunder
                   will raise warning on potential issues. For example if a Job is taking more than
                   an expected amount of time to execute, Thunder will report a warning. This is an
                   extension to the Tracing features.

* [CLEANUP] Seems there is still some old stiff in the CMake files.

---------

Co-authored-by: MFransen69 <39826971+MFransen69@users.noreply.github.com>