[POSTMORTEM] Enhance the dumping of information as required by the Thunder Crash Handler POC [RDK-40121]
From a Thunder perspective this is a usefull part of this POC.
Self-healing, as also explained in the meetings, is an undesirable feature as it break more than it solves.
Recovering in a process is unwanted because:
1) Thunder has the concept of out-0of-process. This is for instable or security sensitive plugins.
   Use this feature if you cannot solve the root-cause, develop stable plugins. A process restart resolves all following objections that come with the described self-healing feature.

2) Attempting to self heal the WPEFramework process will push the cause untrackable issues later on because:
   - memory leakages will occure for objects that have been alllocated (but are not freed due to the self healing)
   - synchronisation object might casue deadlock later on (very hard to track the root-cause than) because the self-healing process can not revert locked synchronisation objects.
   - Files might get corrupted as they might have partially written, Self healing can not close these and recover these.
   - State changes on which other components might rely are never send and thus other components might assume the incorrect state as the change might have occured but the notification was not yet end.

Self healing might suggets fast recovery but as explained in the call, the deadlocking mutex might frustrate the end0user even more as the system becomes than irresponsive for a undefined time.

Fault finding with the spamming RDK Service plugins becomes a nightmare as you need to first check if a self-healing process occured to place the latest crash in the right perspective....

So self healing is a bad idea!