[SECURITY] Add a framework to enable a COMRPCServer per plugin. (#1321)
For security concerns, it is preferred to limit the access over COMRPC to the Thunder global communicator port.
It is preferred to get a dedicated connector per plugin and in such a way limit the access to the plugins by
limiting the access to only that plugin. This way, using containers, access in the container can be opened up
to only a specific subsets of plugins. This can be done in conrainers, through domain sockets, file access/exposure
of the domain socket in the container, or punching through a listening TCP socket using IPTables.

This feature is included in the Thunder framework and needs no dedicated support/alterations in the plugin or plugin
code. If in the main section of the plugin configuration a field "communicator" (as for the main thunder app config)
is declared, that will trigger the opening of a dedicated COMRPC socket for that connector.

If the field value of the "communicator" has a forward slash in the name, e.g. "/communicator|0777", Thunder will open a
domain socket at <volatilepath>/<callsign>/communicator for listening for incoming COMRPC requests (Acquire) specific
for this plugin (no programatical way to reach another plugin)
If the field value of the "communicator" has no forward slash in the name, it will be interpreted as a TCP socket. The
expectation is that the field is organised as "<IP Address>:<portnumber>", "localhost:<portnumber>" or "<hostname>:portnumber"

It goes without saying that this port is only open if the plugin is active. If it is deactivated the listening and all
client sockets are closed!