[Actions] New workflows for documentation preview
Previous mechanism would not work for PRs coming from forked repos
due to security. Create seperate workflows for build/deploy instead. This
is more complex, but gets around the security issues without resorting
to the dangerous pull_request_target workflow:
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/

On PR open, trigger a documentation build & upload site as an artifact.
Then fire an event to run the deployment workflow to publish the site
and leave a comment on the PR

On PR close, trigger a cleanup workflow to remove the preview PR