[REDIRECT] Add a template class to redirect stdout/stderr.
The template has been tested under windows (crappy OS) and should be oke for Linux.
The template will be instantiated to redirect the STDOUT and the STDERR into a non-functional
message stream to be routed to the MessageControl Plugin.
From there the STDOUT/STDERR streams can be routed to any desired destination.