[DANGLING] Decouple Dangling reporting from the Communication thread. (#1366)
The reporting of dangling objects was triggered by the Close of a communication channel. The
communication thread triggeres this Close and thus the reporting of the Dangling pointers is
done on the communication thread.

Stephen Foulds created a test app to validate the Dangling Pointer notification (see
https://github.com/TeknoVenus/ComRpcDanglingTest) this test case forwards the handling of
the dangling notification to an out-of-process part of the plugin. This was causing a
deadlock as the Dangling pointer was reported (and thus handled) on the communication
thread and now this communication thread is needed to "send" and "receive" the COMRPC
messages. However that will not work as the communication thread is now *blocked* in
the COMRPC invoke.

With this PR, the application  is working correctly.

Tested it under Windows, updated Stephen's example app above to run on Windows :-)
See: https://github.com/pwielders/ComRpcDanglingTest