[SECURITY] Check the basic messages for validity (announce/invoke) before passing on. (#1383)
There was no validity check on the size of the message of a COMRPC messages before it was
passed on for further parsing. This meant that if a message was to short, in debug, the
ASSSERT of the further down the line parsing, would fail as the message did not have the
proper length to extract certain pieces of information from it.

These issues, and more might follow, are the result of QA testing with malformed COMRPC
messages.