Development/metrol 784 (#1356)
* [JSONRPC] Non-happy day scenarios on JSONRPC should return an error if possible.

In case the JSONRPC message reeived is malformatted, Thunder should try to return
a JSONRPC error on malformatted JSON message in stead of:
1) Not responding at all.
2) Send an Web Error status even if the ID was sxtracted from the message.

These fixes are related to: METROL-802/METROL-803

* [CONFORMANCE] Use the dedicated JSONRPC error code for incorrect parsing.

* [JSONRPC] Also rubbish after an opaque string should be considered info.

Starting or ending of an opaque structure should *only* contain white spaces. Anything
else should be considered something to be parsed..

* [JSON] Fixes in the JSON parsing to detect issues (added bogus data) earlier.

In the mean time also prevented quoted areas from being accounted in the cope counting and in case we are
deserializing in opaque areas drop the white space for further processing and storage.

* [WEBSOCKET] Messages coming in over websockets could be corrupted as well..

Checking the validity of the JSONRPC messages coming in over the websockets. If incorrect, rturn an error if possible.

* [JSONRPC] Add to each and every JSONRPC register the possibility to check the parameters for parsing errors.

* [Core/JSON] : Produce an error on non-escaped control characters.

* [FIXES] After testing several issues fixed.

* [JSON] Final fix... Finally :-)

* [CLOSE] Report on JSONRPC Notification on HTTP.

* [Core/JSON]: METROL-784.

Empty (opaque) strings should be quoted if enabled.

---------

Co-authored-by: Pierre Wielders <pierre@wielders.net>
Co-authored-by: MFransen69 <39826971+MFransen69@users.noreply.github.com>