[WINDOWS] Make sure warning reporting is not enabled from the core project file.
As warning reporting is optional, I think it is good to enable WarningReporitng builds from the outside. If so the build (by default) does not get
to bulky (lean and mean) and if you want WarningReproting just pass the flag from the ourside (if possible :-).

Also the MessageDispatcher.h is nolonger part of this deplyment, lets kill it :-)