[SWITCHES] Add switches to the config to override flushing and redirecting of stderr and stdout.
In the Thunder confis, in the messaging field, you can now set the field "flush", "stdout" and "stderr" all booleans.
If flush is set to true, all traces foing into the trace system are also "flushed" to the stdout!
If stdout is set false, the stdout will not be redirected to the messagings system!
if stderr is set to false the stderr will not be redirected to the messaging system!