[WINDOWS] The windows implementation of the std::vector is different than the one from Linux.... (#1500)
Changed in a recent pull request the std::list<IResource*> to std::vector<IResource*> as the latter seesm to be more
resource constrained. Tested it on linu and was working fine. On windows!
The root-casue analyses showed that if you "add" an element to the vector, the iterator that you might have
on the vector is invalid after the addition of the interface at the end....
Did not further investigate, just rolled back for windows the std::vector -> std::list again :-)

Co-authored-by: Mateusz Daniluk <121170681+VeithMetro@users.noreply.github.com>