[LIBRARIES] Refernce counted Library COMRPC objects.
If COMRPC objects are created through the C interface, the object itself should hold on to a reference of the library
it is created from this way, the library will not be unloaded as long as there are still COMRPC object living which
reside in this SO. This PR makes sure that any instance instantiated through Core::ServiceType<COMRPCOBJECTIMPL> will
also grab a refenrence onto the SO it is located in. On deletion (Last reference release of the COMRPC object) this
reference on the SO will be dropped as well!