Update IUnknown.h
With the new functionality, the Channel can be invalid if the proxy is not dropped at the moment the channel closes. The proxies are now maintained for reporting (Proxies call on the controller) for leaking proxies. However the reference to the associated channel is dropped at the moment it is closed.