[DEADLOCK] Resolving the ABBA deadlock between the RPC::Administrator and the ProxyStub::UnknownProxy!
The essence is that the calls to RPC::Administrator from within the ProxyStub::UnknownProxy are now always
outside of the administration lock required by the ProxyStub::UnknownProxy. This is possible as the only
member variable that is fluctuating and could cause an issue is the _channel. However, resetting the
_channel in the ProxyStub::UnknownProxy is only possible through the Invalidate. The Invalidate can only
be called by the RPC::Administrator, whois now usin it's own lock in case the Invalidate is called.
This way using/checking the _channel member frm ProxyStub::UnknownProxy is now safe from an RPC::Administrator
point of view.
But the proof of the pudding is in the eating, time to start eating :-)