[OUT-OF-PROCESS] Move the Plugin, as a whole out-of-process.
On request of Sky Labs Aalborg, we are looking at possibilities to run any plugin, also plugins not coded
for out of process, out-of-process to make sure that memory corruption created by such is plugin is only
in its own process space and nolonger in the Thunder (WPEFramework) process space.

Focus for this feature are mainly plugins created in the RDKService that are not yet prepared to run out of
process but are only exposing a JSONRPC to the outside world.

The idea is to pass a root config object on plugin level, not on configuration level. If such a root config
object is set on plugin level (keyword also "root"), the Framework will start/load the plugin out of process
in the ThunderPlugin application process space. The request made (through the regular code) is to return a
IPlugin (one of the perls of working interface based all along ;-) ) Now if we retrieve the IPlugin from the
module loaded in the ThunderPlugin process space, Thunder framework will continue with this IPlugin as if it
where created locally and thus the IDispatcher (For JSONRPC dispatching) are extracted from the IPlugin for
handling the JSONRPC over COMRPC (feature available as of R4.4).

The Initialize is now also being forwarded through the IPlugin to the instance that is running in the
ThunderPlugin process (out-of-process hosting). This means that Initialization of the plugin is only done in
the ThunderPlugin process and nolonger in the Thunder (WPEFramework) process so this might lead to unexpected
situations. To be found during testing :-)

Another point of attention/testing is the feedback of JSONRPC events. More commits might follow to get that
properly working.