[OUT-OF-PROCESS] Tested it succefully, more testing needed..
Issues:
IDispatcher needs to be intialized and deinitialized, added the ILocalDispatcher methods to the
IDispatcher, so the JSONRPC class can also be intialized (Attach) and deinitialized (detach) in
the other process space.

Crash due to a full Deactivate in case an Activate fails and the new behaviour (Deinitialize called
after any faiing Initialize) was assuming JSONRPC stuff was set, which was not :-) Now aonly call
Deinitialize in stead of a full Deactivate.

Decoupled notification of the channels that are opened and closed. As the closing of the channel
was detected by the ResourceMonitor thread, this thread is used to also iterate over the subscribers
for notifiation of channel opening and closing. As JSONRPC classes are one of the interested parties
to this event, the event now had to travel across a COMRPC boundary, in case the JSONRPC is running
out-of-process. A COMRPC call is a no go on the ResourceMonitor thread since that will deadlock ;-)

The registration from the IConnectionServer::INotification for the out-of-process plugin was triggered
by the Attach of the IDispatcher, however since this registration actively pushes all current open
channels to the subscriber we ended up in a deadlock:
COMRPC classical deadlock:
Thunder -> (IDispatcher::Attach)   ThunderPlugin --+
                                                   |
 +-------   IShell::Register(sink) <---------------+
 |
 +------> INotification::Open(id) [Oops, in and outbound on COMRPC are already in progress!!!!]

Solved by pushing the sink back on the Attach (and later Detach for symetry) and have Thunder Register
this sink on behalf of the ouot of process plugin !

Tested sofar calling JSONRPC method on the out-of-process plugin (JSONRPCPlugin in examples of ThunderNanoServices,
which worked flawless and did receive already notification that are send by default to the system (inbound
events)

More testing needs to be done!